Move collections e2e tests from product collection file to its own file
